requirements.txt
requests
python-interface